Refactor accessrequests to its own package
