+++ internal/auth/accessrequest/handlers.go
@@ -9,6 +9,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/auth/userpasswd"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
+	"github.com/sourcegraph/sourcegraph/internal/database/accessrequests"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 
 	"github.com/sourcegraph/sourcegraph/internal/database"
@@ -63,7 +64,9 @@ func handleRequestAccess(logger log.Logger, db database.DB, w http.ResponseWrite
 		Email:          data.Email,
 		AdditionalInfo: data.AdditionalInfo,
 	}
-	_, err := db.AccessRequests().Create(r.Context(), &accessRequest)
+	accessRequestStore := accessrequests.With(db.GetStore(), logger)
+
+	_, err := accessRequestStore.Create(r.Context(), &accessRequest)
 	if err == nil {
 		w.WriteHeader(http.StatusCreated)
 		if err = usagestats.LogBackendEvent(db, actor.FromContext(r.Context()).UID, deviceid.FromContext(r.Context()), "CreateAccessRequestSucceeded", nil, nil, featureflag.GetEvaluatedFlagSet(r.Context()), nil); err != nil {
@@ -72,7 +75,7 @@ func handleRequestAccess(logger log.Logger, db database.DB, w http.ResponseWrite
 		return
 	}
 	logger.Error("Error in access request.", log.String("email", data.Email), log.String("name", data.Name), log.Error(err))
-	if database.IsAccessRequestUserWithEmailExists(err) || database.IsAccessRequestWithEmailExists(err) {
+	if accessrequests.IsAccessRequestUserWithEmailExists(err) || accessrequests.IsAccessRequestWithEmailExists(err) {
 		// 🚨 SECURITY: We don't show an error message when the user or access request with the same e-mail address exists
 		// as to not leak the existence of a given e-mail address in the database.
 		w.WriteHeader(http.StatusCreated)
