+++ internal/database/accessrequests/access_requests.go
@@ -1,4 +1,4 @@
-package database
+package accessrequests
 
 import (
 	"context"
@@ -8,6 +8,7 @@ import (
 	"github.com/keegancsmith/sqlf"
 	"github.com/sourcegraph/log"
 
+	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/types"
@@ -80,7 +81,7 @@ type AccessRequestStore interface {
 	GetByID(context.Context, int32) (*types.AccessRequest, error)
 	GetByEmail(context.Context, string) (*types.AccessRequest, error)
 	Count(context.Context, *AccessRequestsFilterArgs) (int, error)
-	List(context.Context, *AccessRequestsFilterArgs, *PaginationArgs) (_ []*types.AccessRequest, err error)
+	List(context.Context, *AccessRequestsFilterArgs, *database.PaginationArgs) (_ []*types.AccessRequest, err error)
 	WithTransact(context.Context, func(AccessRequestStore) error) error
 	Done(error) error
 }
@@ -90,8 +91,8 @@ type accessRequestStore struct {
 	logger log.Logger
 }
 
-// AccessRequestsWith instantiates and returns a new accessRequestStore using the other store handle.
-func AccessRequestsWith(other basestore.ShareableStore, logger log.Logger) AccessRequestStore {
+// With instantiates and returns a new accessRequestStore using the other store handle.
+func With(other basestore.ShareableStore, logger log.Logger) AccessRequestStore {
 	return &accessRequestStore{Store: basestore.NewWithHandle(other.Handle()), logger: logger}
 }
 
@@ -184,7 +185,6 @@ func (s *accessRequestStore) Create(ctx context.Context, accessRequest *types.Ac
 func (s *accessRequestStore) GetByID(ctx context.Context, id int32) (*types.AccessRequest, error) {
 	row := s.QueryRow(ctx, sqlf.Sprintf("SELECT %s FROM access_requests WHERE id = %s", sqlf.Join(accessRequestColumns, ","), id))
 	node, err := scanAccessRequest(row)
-
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return nil, &ErrAccessRequestNotFound{ID: id}
@@ -198,7 +198,6 @@ func (s *accessRequestStore) GetByID(ctx context.Context, id int32) (*types.Acce
 func (s *accessRequestStore) GetByEmail(ctx context.Context, email string) (*types.AccessRequest, error) {
 	row := s.QueryRow(ctx, sqlf.Sprintf("SELECT %s FROM access_requests WHERE email = %s", sqlf.Join(accessRequestColumns, ","), email))
 	node, err := scanAccessRequest(row)
-
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return nil, &ErrAccessRequestNotFound{Email: email}
@@ -212,7 +211,6 @@ func (s *accessRequestStore) GetByEmail(ctx context.Context, email string) (*typ
 func (s *accessRequestStore) Update(ctx context.Context, accessRequest *types.AccessRequest) (*types.AccessRequest, error) {
 	q := sqlf.Sprintf(accessRequestUpdateQuery, accessRequest.Status, *accessRequest.DecisionByUserID, accessRequest.ID, sqlf.Join(accessRequestColumns, ","))
 	updated, err := scanAccessRequest(s.QueryRow(ctx, q))
-
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, &ErrAccessRequestNotFound{ID: accessRequest.ID}
@@ -228,13 +226,13 @@ func (s *accessRequestStore) Count(ctx context.Context, fArgs *AccessRequestsFil
 	return basestore.ScanInt(s.QueryRow(ctx, q))
 }
 
-func (s *accessRequestStore) List(ctx context.Context, fArgs *AccessRequestsFilterArgs, pArgs *PaginationArgs) ([]*types.AccessRequest, error) {
+func (s *accessRequestStore) List(ctx context.Context, fArgs *AccessRequestsFilterArgs, pArgs *database.PaginationArgs) ([]*types.AccessRequest, error) {
 	if fArgs == nil {
 		fArgs = &AccessRequestsFilterArgs{}
 	}
 	where := fArgs.SQL()
 	if pArgs == nil {
-		pArgs = &PaginationArgs{}
+		pArgs = &database.PaginationArgs{}
 	}
 	p := pArgs.SQL()
