+++ internal/database/database.go
@@ -22,7 +22,8 @@ type DB interface {
 	dbutil.DB
 	basestore.ShareableStore
 
-	AccessRequests() AccessRequestStore
+	GetStore() basestore.ShareableStore
+
 	AccessTokens() AccessTokenStore
 	Authz() AuthzStore
 	BitbucketProjectPermissions() BitbucketProjectPermissionsStore
@@ -134,10 +135,6 @@ func (d *db) Done(err error) error {
 
 func (d *db) AccessTokens() AccessTokenStore {
 	return AccessTokensWith(d.Store, d.logger.Scoped("AccessTokenStore", ""))
-}
-
-func (d *db) AccessRequests() AccessRequestStore {
-	return AccessRequestsWith(d.Store, d.logger.Scoped("AccessRequestStore", ""))
 }
 
 func (d *db) BitbucketProjectPermissions() BitbucketProjectPermissionsStore {
@@ -369,4 +366,8 @@ func (d *db) AssignedTeams() AssignedTeamsStore {
 
 func (d *db) OwnSignalConfigurations() SignalConfigurationStore {
 	return SignalConfigurationStoreWith(d.Store)
+}
+
+func (d *db) GetStore() basestore.ShareableStore {
+	return d.Store
 }
